requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.8.0rc2
dj-rest-auth==5.0.2
Django==4.2.9
django-allauth==0.60.0
django-cors-headers==4.3.1
django-rest-registration==0.8.3
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
idna==3.6
mysql-connector-python==8.2.0
mysqlclient==2.2.1
oauthlib==3.2.2
pillow==10.2.0
protobuf==4.21.12
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
typing-extensions==4.9.0
urllib3==2.1.0
whitenoise==6.6.0
